Move holiday-stop credit calculation
