+++ handlers/holiday-stop-processor/src/main/scala/com/gu/holidaystopprocessor/ProcessResult.scala
@@ -1,5 +1,6 @@
 package com.gu.holidaystopprocessor
 
+import com.gu.holiday_stops.{HolidayStop, OverallFailure, ZuoraHolidayWriteError}
 import com.typesafe.scalalogging.LazyLogging
 
 case class ProcessResult(holidayStopsToApply: List[HolidayStop], holidayStopResults: List[Either[ZuoraHolidayWriteError, HolidayStopResponse]], resultsToExport: List[HolidayStopResponse], overallFailure: Option[OverallFailure])
