+++ handlers/holiday-stop-processor/src/test/scala/com/gu/holidaystopprocessor/SubscriptionTest.scala
@@ -2,6 +2,7 @@ package com.gu.holidaystopprocessor
 
 import java.time.LocalDate
 
+import com.gu.holiday_stops.RatePlanCharge
 import org.scalatest.{FlatSpec, Matchers, OptionValues}
 
 class SubscriptionTest extends FlatSpec with Matchers with OptionValues {
