+++ lib/holiday-stops/src/main/scala/com/gu/holiday_stops/AccessToken.scala
@@ -1,3 +1,3 @@
-package com.gu.holidaystopprocessor
+package com.gu.holiday_stops
 
 case class AccessToken(access_token: String)
