+++ lib/holiday-stops/src/main/scala/com/gu/holiday_stops/HolidayCreditUpdate.scala
@@ -1,6 +1,8 @@
-package com.gu.holidaystopprocessor
+package com.gu.holiday_stops
 
 import java.time.LocalDate
+
+import com.gu.holiday_stops
 
 case class HolidayCreditUpdate(
   currentTerm: Option[Int],
