+++ lib/holiday-stops/src/main/scala/com/gu/holiday_stops/HolidayStop.scala
@@ -1,4 +1,4 @@
-package com.gu.holidaystopprocessor
+package com.gu.holiday_stops
 
 import java.time.LocalDate
