+++ lib/holiday-stops/src/main/scala/com/gu/holiday_stops/Subscription.scala
@@ -1,4 +1,4 @@
-package com.gu.holidaystopprocessor
+package com.gu.holiday_stops
 
 import java.time.LocalDate
 
@@ -12,32 +12,22 @@ case class Subscription(
   ratePlans: List[RatePlan]
 ) {
 
-  val originalRatePlanCharge: Option[RatePlanCharge] = {
-    val chronologicallyOrderedRatePlans = ratePlans.sortBy { plan =>
-      plan.ratePlanCharges.map(_.effectiveStartDate.toString).headOption.getOrElse("")
-    }
-    for {
-      ratePlan <- chronologicallyOrderedRatePlans.headOption
-      charge <- ratePlan.ratePlanCharges.headOption
-    } yield charge
-  }
-
   def ratePlanCharge(stop: HolidayStop): Option[RatePlanCharge] = {
 
     def isMatchingPlan(plan: RatePlan): Boolean = plan.productName == "Discounts"
 
     def isMatchingCharge(charge: RatePlanCharge): Boolean =
       charge.name == "Holiday Credit" &&
-        charge.HolidayStart__c.exists { start =>
-          start.isEqual(stop.stoppedPublicationDate) || start.isBefore(stop.stoppedPublicationDate)
-        } &&
-        charge.HolidayEnd__c.exists { end =>
-          end.isEqual(stop.stoppedPublicationDate) || end.isAfter(stop.stoppedPublicationDate)
-        }
+      charge.HolidayStart__c.exists { start =>
+        start.isEqual(stop.stoppedPublicationDate) || start.isBefore(stop.stoppedPublicationDate)
+      } &&
+      charge.HolidayEnd__c.exists { end =>
+        end.isEqual(stop.stoppedPublicationDate) || end.isAfter(stop.stoppedPublicationDate)
+      }
 
     val charges = for {
       plan <- ratePlans if isMatchingPlan(plan)
-      charge <- plan.ratePlanCharges.find(isMatchingCharge)
+        charge <- plan.ratePlanCharges.find(isMatchingCharge)
     } yield charge
     charges.headOption
   }
@@ -63,5 +53,3 @@ case class RatePlanCharge(
   HolidayEnd__c: Option[LocalDate],
   processedThroughDate: Option[LocalDate],
 )
-
-
