+++ lib/holiday-stops/src/main/scala/com/gu/holiday_stops/ZuoraStatusResponse.scala
@@ -1,4 +1,4 @@
-package com.gu.holidaystopprocessor
+package com.gu.holiday_stops
 
 case class ZuoraStatusResponse(
   success: Boolean,
